requirements.txt
garminconnect
pandas
matplotlib
numpy
pandera
python-dotenv
dlt[duckdb]
duckdb
streamlit